Web scraping and dashboard
This project is about web scraping the data from different E-commerce websites and comparing the price between them. Here I have used selenium in Python to scrap the data for Different websites and created a price comparison dashboard using Power BI. If the code has error kindly check the css.selector in the code and change it accordingly.